backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
requests>=2.28.2

# Cryptography and security
cryptography>=40.0.1
pyjwt>=2.6.0

# Data processing
numpy>=1.24.2
pandas>=2.0.0
beautifulsoup4>=4.12.2

# Testing
pytest>=7.3.1

# OpenAI integration
openai>=0.27.4